Move JSON update test mixin to helpers
